Uploaded static website
Final assignment for Web Programming 1 course: a fully-functional static website made to sell a fictional product, in this case weapons, armours, and magical items. Includes multiple pages, such as a a home page, 3 different store pages, a video page, and a contact page, and is both mobile and desktop friendly.